+++ Model/user.go
@@ -1,6 +1,7 @@
 package Model
 
 import (
+	uuid "github.com/satori/go.uuid"
 	"go-study/Library"
 	"go-study/Library/Gorm"
 	"gorm.io/gorm"
@@ -15,6 +16,7 @@ var tableName = "user"
 //
 type User struct {
 	ID        uint64        `gorm:"column:id;primaryKey;autoIncrement" json:"id"`
+	Uuid      uuid.UUID     `gorm:"column:uuid;" json:"uuid"`
 	Name      string        `gorm:"column:name;" json:"name"`
 	CreatedAt *Library.Time `gorm:"column:created_at;" json:"created_at"`
 	UpdatedAt *Library.Time `gorm:"column:updated_at;" json:"updated_at"`
