+++ Repository/User/user.go
@@ -16,3 +16,15 @@ func FindById(id uint64) (userInfo Model.User) {
 
 	return
 }
+
+//
+// Create
+// @Description: 新增用户方法
+// @param user
+// @return uint64
+//
+func Create(user Model.User) uint64 {
+	Model.UserModel().Create(&user)
+
+	return user.ID
+}
